refactor(fe): separate CreateContestForm from create page
