+++ apps/frontend/app/admin/contest/create/page.tsx
@@ -20,24 +20,13 @@ import {
   DialogTitle
 } from '@/components/shadcn/dialog'
 import { ScrollArea } from '@/components/shadcn/scroll-area'
-import {
-  CREATE_CONTEST,
-  IMPORT_PROBLEMS_TO_CONTEST
-} from '@/graphql/contest/mutations'
-import { UPDATE_CONTEST_PROBLEMS_ORDER } from '@/graphql/problem/mutations'
-import { useMutation } from '@apollo/client'
-import type { CreateContestInput } from '@generated/graphql'
-import { valibotResolver } from '@hookform/resolvers/valibot'
 import { ErrorBoundary } from '@suspensive/react'
 import { PlusCircleIcon } from 'lucide-react'
 import Link from 'next/link'
-import { useRouter } from 'next/navigation'
-import { useState, useRef, Suspense } from 'react'
-import { useForm, FormProvider } from 'react-hook-form'
+import { useState, Suspense } from 'react'
 import { FaAngleLeft } from 'react-icons/fa6'
 import { IoMdCheckmarkCircleOutline } from 'react-icons/io'
-import { toast } from 'sonner'
-import { useConfirmNavigation } from '../../_components/ConfirmNavigation'
+import ConfirmNavigation from '../../_components/ConfirmNavigation'
 import DescriptionForm from '../../_components/DescriptionForm'
 import FormSection from '../../_components/FormSection'
 import SwitchField from '../../_components/SwitchField'
@@ -49,115 +38,30 @@ import {
   ImportProblemTableFallback
 } from '../_components/ImportProblemTable'
 import TimeForm from '../_components/TimeForm'
-import { type ContestProblem, createSchema } from '../_libs/schemas'
+import type { ContestProblem } from '../_libs/schemas'
+import CreateContestForm from './_components/CreateContestForm'
 
 export default function Page() {
   const [problems, setProblems] = useState<ContestProblem[]>([])
-  const [isCreating, setIsCreating] = useState<boolean>(false)
-  const [showImportDialog, setShowImportDialog] = useState<boolean>(false)
-
-  const shouldSkipWarning = useRef(false)
-  const router = useRouter()
-
-  useConfirmNavigation(shouldSkipWarning)
-
-  const methods = useForm<CreateContestInput>({
-    resolver: valibotResolver(createSchema),
-    defaultValues: {
-      invitationCode: null,
-      isRankVisible: true,
-      isVisible: true,
-      enableCopyPaste: false,
-      isJudgeResultVisible: false
-    }
-  })
-
-  const { handleSubmit } = methods
-
-  const [createContest, { error }] = useMutation(CREATE_CONTEST)
-  const [importProblemsToContest] = useMutation(IMPORT_PROBLEMS_TO_CONTEST)
-  const [updateContestProblemsOrder] = useMutation(
-    UPDATE_CONTEST_PROBLEMS_ORDER
-  )
-
-  const isSubmittable = (input: CreateContestInput) => {
-    if (input.startTime >= input.endTime) {
-      toast.error('Start time must be less than end time')
-      return
-    }
-
-    if (
-      new Set(problems.map((problem) => problem.order)).size !== problems.length
-    ) {
-      toast.error('Duplicate problem order found')
-      return
-    }
-    onSubmit()
-  }
-
-  const onSubmit = async () => {
-    const input = methods.getValues()
-    setIsCreating(true)
-
-    const { data } = await createContest({
-      variables: {
-        groupId: 1,
-        input
-      }
-    })
-
-    const contestId = Number(data?.createContest.id)
-    if (error) {
-      toast.error('Failed to create contest')
-      setIsCreating(false)
-      return
-    }
-    await importProblemsToContest({
-      variables: {
-        groupId: 1,
-        contestId,
-        problemIdsWithScore: problems.map((problem) => {
-          return {
-            problemId: problem.id,
-            score: problem.score
-          }
-        })
-      }
-    })
-    const orderArray = problems
-      .sort((a, b) => a.order - b.order)
-      .map((problem) => problem.id)
-    await updateContestProblemsOrder({
-      variables: {
-        groupId: 1,
-        contestId,
-        orders: orderArray
-      }
-    })
-
-    shouldSkipWarning.current = true
-    toast.success('Contest created successfully')
-    router.push('/admin/contest')
-    router.refresh()
-  }
+  const [isCreating, setIsCreating] = useState(false)
+  const [showImportDialog, setShowImportDialog] = useState(false)
 
   return (
-    <ScrollArea className="w-full">
-      <main className="flex flex-col gap-6 px-20 py-16">
-        <div className="flex items-center gap-4">
-          <Link href="/admin/contest">
-            <FaAngleLeft className="h-12" />
-          </Link>
-          <span className="text-4xl font-bold">Create Contest</span>
-        </div>
-        <form
-          onSubmit={handleSubmit(isSubmittable)}
-          className="flex w-[760px] flex-col gap-6"
-        >
-          <FormProvider {...methods}>
+    <ConfirmNavigation>
+      <ScrollArea className="w-full">
+        <main className="flex flex-col gap-6 px-20 py-16">
+          <div className="flex items-center gap-4">
+            <Link href="/admin/contest">
+              <FaAngleLeft className="h-12" />
+            </Link>
+            <span className="text-4xl font-bold">Create Contest</span>
+          </div>
+
+          <CreateContestForm problems={problems} setIsCreating={setIsCreating}>
             <FormSection title="Title">
               <TitleForm placeholder="Name your contest" />
             </FormSection>
+
             <div className="flex gap-6">
               <FormSection title="Start Time">
                 <TimeForm name="startTime" />
@@ -166,17 +70,21 @@ export default function Page() {
                 <TimeForm name="endTime" />
               </FormSection>
             </div>
+
             <FormSection title="Description">
               <DescriptionForm name="description" />
             </FormSection>
+
             <SwitchField
               name="enableCopyPaste"
               title="Enable participants Copy/Pasting"
             />
+
             <SwitchField
               name="isJudgeResultVisible"
               title="Reveal scores to participants"
             />
+
             <SwitchField
               name="invitationCode"
               title="Invitation Code"
@@ -254,6 +162,7 @@ export default function Page() {
                 disableInput={false}
               />
             </div>
+
             <Button
               type="submit"
               className="flex h-[36px] w-[100px] items-center gap-2 px-0"
@@ -262,9 +171,9 @@ export default function Page() {
               <IoMdCheckmarkCircleOutline fontSize={20} />
               <div className="mb-[2px] text-base">Create</div>
             </Button>
-          </FormProvider>
-        </form>
-      </main>
-    </ScrollArea>
+          </CreateContestForm>
+        </main>
+      </ScrollArea>
+    </ConfirmNavigation>
   )
 }
